feat: move macros into their own file
